+++ src/daemon/src/com/aerofs/daemon/transport/xmpp/jingle/SignalThread.java
@@ -171,9 +171,11 @@ public class SignalThread extends java.lang.Thread implements IDumpStatMisc
         // We avoid resolving the hostname ourselves and let
         // SMACK do the DNS query on its thread.
         InetSocketAddress address = Param.xmppAddress();
+        // TODO (WW) XmppMain() should use int rathe than short as the datatype of jingleRelayPort
+        // as Java's unsigned short may overflow on big port numbers.
         _main = new XmppMain(address.getHostName(), address.getPort(),
                 true, _jidSelf, XMPPServerConnection.shaedXMPP(),
-                L.get().jingleRelayHost(), L.get().jingleRelayPort(),
+                L.get().jingleRelayHost(), (short) L.get().jingleRelayPort(),
                 ljlogpathutf8);
 
         l.debug("st: created xmppmain");
